Combine cloud and network refreshes
